OSD-10224 Refactor to introduce k8s dependency
