+++ pkg/handlers/paths.go
@@ -1,0 +1,12 @@
+package handlers
+
+const (
+	// Set ready probe path
+	ReadyzPath = "/readyz"
+
+	// Set live probe path
+	LivezPath = "/livez"
+
+	// Set webhook receiver path
+	WebhookReceiverPath = "/alertmanager-receiver"
+)
